backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.2.3
mysql-connector-python==8.2.0
PyJWT==2.8.0
